+++ piceli/__init__.py
@@ -22,12 +22,16 @@ _APP_EXPORTS = frozenset(
         "MemoryVolume",
         "Mount",
         "NetworkPolicy",
+        "PodDefaults",
         "Probe",
         "Resources",
+        "Rule",
         "Secret",
         "SecretKey",
         "SecretVolume",
+        "Security",
         "Service",
+        "ServiceAccount",
         "ServicePort",
     }
 )
@@ -71,12 +75,16 @@ if TYPE_CHECKING:
         MemoryVolume,
         Mount,
         NetworkPolicy,
+        PodDefaults,
         Probe,
         Resources,
+        Rule,
         Secret,
         SecretKey,
         SecretVolume,
+        Security,
         Service,
+        ServiceAccount,
         ServicePort,
     )
     from piceli.checks import Checks  # noqa: F401